refactor(web): route standard — /signin only auth route, app IA at /dashboard
- rename /login → /signin (308 redirect kept at /login); /app → /dashboard
- quarantine legacy /dashboard tree + NavBar/MenuBar/ProtectedRoute → src/legacy
- home nav swapped to W1 MarketingNav (legacy NavBar linked retired /login)
- signin.spec.ts covers signin → dashboard loop + /login redirect
- docs/web-implementation.md route map, stage rows, TEST_MODE, §8 updated
