+++ web/e2e/signin.spec.ts
@@ -1,14 +1,14 @@
 import { expect, test } from "@playwright/test";
 
 /**
- * TEST_MODE smoke: the X-1 flow — /login shows exactly one Google CTA;
- * clicking it lands on the new IA root (/app) with data served by the
- * in-app mock server.
+ * TEST_MODE smoke: the X-1 flow — /signin shows exactly one Google CTA;
+ * clicking it lands on the app root (/dashboard, route standard) with data
+ * served by the in-app mock server.
  */
-test("login → app home via the single Google CTA", async ({ page }) => {
-  await page.goto("/login");
+test("signin → dashboard home via the single Google CTA", async ({ page }) => {
+  await page.goto("/signin");
 
-  const screen = page.getByTestId("login-screen");
+  const screen = page.getByTestId("signin-screen");
   await expect(screen).toBeVisible();
   await expect(screen.getByRole("heading", { name: "Sign in to Upstat" })).toBeVisible();
 
@@ -19,12 +19,18 @@ test("login → app home via the single Google CTA", async ({ page }) => {
   await expect(screen.getByRole("button")).toHaveCount(1);
 
   await cta.click();
-  await page.waitForURL("**/app");
+  await page.waitForURL("**/dashboard");
 
-  const home = page.getByTestId("app-home");
+  const home = page.getByTestId("dashboard-home");
   await expect(home).toBeVisible();
   // Seeded org narrative served by the mock server.
   await expect(home.getByText("Upstat · Africa/Lagos")).toBeVisible();
+});
+
+test("/login permanently redirects to /signin (no 404 for old links)", async ({ request }) => {
+  const response = await request.get("/login", { maxRedirects: 0 });
+  expect(response.status()).toBe(308);
+  expect(response.headers()["location"]).toContain("/signin");
 });
 
 test("mock server serves the seeded narrative", async ({ request }) => {
