+++ web/src/app/dashboard/page.tsx
@@ -1,22 +1,34 @@
 "use client";
 
-import { useDateRange } from "@/components/hooks/use-date-range";
-import DateRangeFilter from "@/components/pages/dashboard/date-range-filter";
-import StatsOverview from "@/components/pages/dashboard/stats-overview";
-import TotalUsersChart from "@/components/pages/dashboard/total-users-chart";
-import { ChartsRow } from "@/components/pages/dashboard/charts-layout.styles";
+import { useOrgController } from "@/controllers/onboarding";
 
-export default function Dashboard() {
-  const dateRange = useDateRange();
+/**
+ * B1 Home placeholder — the W3 stage builds the real org-health screen
+ * (incident banner, triggered monitors, SLO burn, watched dashboards).
+ * W0 proves the signin → dashboard flow and the controller → repository →
+ * mock server path end to end.
+ */
+export default function DashboardHomePage() {
+  const { data: org, loading } = useOrgController();
 
   return (
-    <div>
-      <DateRangeFilter dateRange={dateRange} />
-      <StatsOverview range={dateRange.selected} />
-
-      <ChartsRow>
-        <TotalUsersChart range={dateRange.selected} />
-       </ChartsRow>
-    </div>
+    <main
+      data-testid="dashboard-home"
+      className="mx-auto flex max-w-[960px] flex-col gap-[var(--space-4)] px-[var(--space-8)] py-[var(--space-12)]"
+    >
+      <h1 className="text-[20px] font-semibold">Home</h1>
+      <p className="text-[13px] leading-[1.45] text-text-2">
+        {loading
+          ? "Loading org…"
+          : org
+            ? `${org.name} · ${org.timezone}`
+            : "No organization yet."}
+      </p>
+      <p className="rounded-(--radius) border border-border bg-bg-elev p-[var(--space-4)] text-[13px] leading-[1.45] text-text-2">
+        The observability pillars land here in W3 (pages.md Part B). This
+        screen exists so W0 can prove sign-in, controllers, and the mock
+        server end to end.
+      </p>
+    </main>
   );
 }
