+++ web/src/legacy/app/dashboard/error/page.tsx
