+++ web/src/legacy/components/layout/nav-bar/nav-bar.styles.ts
