+++ web/src/legacy/components/shared-layouts/menu-bar/menu-bar.styles.ts
